Fixed mentor's remarks in Fruit storage

+++ src/main/java/core/basesyntax/dao/FruitStorage.java
@@ -1,9 +1,8 @@
 package core.basesyntax.dao;
 
-import java.math.BigDecimal;
 import java.util.HashMap;
 import java.util.Map;
 
 public class FruitStorage {
-    public static final Map<String, BigDecimal> storage = new HashMap<>();
+    public static final Map<String, Integer> storage = new HashMap<>();
 }
