+++ src/main/java/core/basesyntax/service/ShopService.java
@@ -1,8 +1,7 @@
 package core.basesyntax.service;
 
 import core.basesyntax.model.Operation;
-import java.math.BigDecimal;
 
 public interface ShopService {
-    void transfer(Operation operation, String fruit, BigDecimal quantity);
+    void transfer(Operation operation, String fruit, Integer quantity);
 }
