+++ src/main/java/core/basesyntax/service/WriterReportToCsv.java
@@ -1,8 +1,7 @@
 package core.basesyntax.service;
 
-import java.math.BigDecimal;
 import java.util.Map;
 
 public interface WriterReportToCsv {
-    void writeReport(Map<String, BigDecimal> storage, String outputFilePath);
+    void writeReport(Map<String, Integer> storage, String outputFilePath);
 }
